+++ llm_thread/tests/__init__.py
@@ -2,6 +2,7 @@ from . import test_process_llm_body
 from . import test_mail_message_store
 from . import test_message_post_escape
 from . import test_thread_create_store
+from . import test_context_stats
 from . import test_llm_thread_tag
 from . import test_llm_thread_search
 from . import test_message_compute_author